+++ app/domain/keywordmatcher.py
@@ -1,5 +1,4 @@
-from typing import List
-from domain import AbstractQuestionMatcher
+from .questionmatcher import AbstractQuestionMatcher
 
 class KeywordMatcher(AbstractQuestionMatcher):
     def __init__(self):
@@ -7,7 +6,7 @@ class KeywordMatcher(AbstractQuestionMatcher):
 
         print("Keyword matcher loaded")
 
-    def getSuggestions(self, question: str) -> List[str]:
+    def getSuggestions(self, question: str) -> list[str]:
         pastQuestions = {"pointer": "How do I use pointers in C?", "code": "How do I code well?"}
         suggestions = []
